consumer refactor before review

+++ src/main/java/by/home/fileConsumer/service/FileService.java
@@ -12,6 +12,9 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
 
+import static by.home.fileConsumer.component.Util.addString;
+import static by.home.fileConsumer.component.Util.validate;
+
 @Service
 public class FileService {
 
@@ -20,20 +23,44 @@ public class FileService {
     @Value("${directory.path}")
     private String pathToDirectory;
 
-    public void createWrite(List<FileTransferModel> listFile) {
+    /**
+     * Parsing list of FileTransferModel to name of file and body.
+     * then createWriteFile saving
+     *
+     * @param listFile list of FileTransferModel
+     */
+    public void saveFile(List<FileTransferModel> listFile) {
         for (FileTransferModel currentFile : listFile) {
             createWriteFile(currentFile.getName(), currentFile.getBody());
         }
     }
 
-    private boolean createWriteFile(String name, byte[] bFile) {
-        Path path = Paths.get(pathToDirectory + name);
+    /**
+     * Create files and filling up
+     *
+     * @param name  file name
+     * @param bFile array of byte
+     */
+    private void createWriteFile(String name, byte[] bFile) {
+        validate(pathToDirectory == null, "directory.not.found");
+        validate(name == null, "file.not create");
+        validate(!Files.isDirectory(Paths.get(pathToDirectory)), "directory.not.found");
         try {
-            Files.createFile(path);
-            Files.write(path, bFile);
+            Path path = Paths.get(pathToDirectory + name);
+            if (Files.exists(path)) {
+                long numberDuplicate = Files.list(path.getParent()).count();
+                String refactorString = addString(name, String.valueOf(numberDuplicate), name.indexOf("."));
+                Files.createFile(Paths.get(pathToDirectory + refactorString));
+                Files.write(Paths.get(pathToDirectory + refactorString), bFile);
+            } else {
+                Files.createFile(path);
+                Files.write(path, bFile);
+            }
         } catch (IOException ex) {
-            System.err.println("File already exists");
+            validate(true, "file.not.create");
+            LOGGER.error(ex.getMessage());
         }
-        return false;
     }
+
+
 }
